replaced *.csproj with new *.xproj
