+++ Properties/AssemblyInfo.cs
@@ -1,0 +1,23 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("RazorPDFCore")]
+[assembly: AssemblyDescription("RazorPDFCore is used to convert PDF documents usign itextshap and razor views")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Al Nyveldt")]
+[assembly: AssemblyProduct("RazorPDFCore")]
+[assembly: AssemblyCopyright("Copyright ©  2016")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+// Setting ComVisible to false makes the types in this assembly not visible
+// to COM components.  If you need to access a type in this assembly from
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("523cd2d5-daa0-479b-9d5c-b133daf48978")]
