--- RazorPDF.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RazorPDF", "RazorPDF\RazorPDF.csproj", "{0EEB1093-C4B8-4830-BC01-2C56E618702E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0EEB1093-C4B8-4830-BC01-2C56E618702E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0EEB1093-C4B8-4830-BC01-2C56E618702E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0EEB1093-C4B8-4830-BC01-2C56E618702E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0EEB1093-C4B8-4830-BC01-2C56E618702E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
