--- RazorPDF/PdfResultExecutor.cs
@@ -1,258 +1,0 @@
-﻿// Copyright 2016 Al Nyveldt - http://nyveldt.com, Ole Koeckemann <ole.k@web.de>
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// 
-using iTextSharp.text;
-using iTextSharp.text.html.simpleparser;
-using iTextSharp.text.pdf;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Internal;
-using Microsoft.AspNetCore.Mvc.ModelBinding;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.AspNetCore.Mvc.ViewEngines;
-using Microsoft.AspNetCore.Mvc.ViewFeatures;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RazorPDFCore
-{
-    public class PdfResultExecutor : ViewExecutor
-    {
-        private const string ActionNameKey = "action";
-
-        /// <summary>
-        /// Creates a new <see cref="ViewResultExecutor"/>.
-        /// </summary>
-        /// <param name="viewOptions">The <see cref="IOptions{MvcViewOptions}"/>.</param>
-        /// <param name="writerFactory">The <see cref="IHttpResponseStreamWriterFactory"/>.</param>
-        /// <param name="viewEngine">The <see cref="ICompositeViewEngine"/>.</param>
-        /// <param name="tempDataFactory">The <see cref="ITempDataDictionaryFactory"/>.</param>
-        /// <param name="diagnosticSource">The <see cref="DiagnosticSource"/>.</param>
-        /// <param name="loggerFactory">The <see cref="ILoggerFactory"/>.</param>
-        /// <param name="modelMetadataProvider">The <see cref="IModelMetadataProvider"/>.</param>
-        public PdfResultExecutor(
-            IOptions<MvcViewOptions> viewOptions,
-            IHttpResponseStreamWriterFactory writerFactory,
-            ICompositeViewEngine viewEngine,
-            ITempDataDictionaryFactory tempDataFactory,
-            DiagnosticSource diagnosticSource,
-            ILoggerFactory loggerFactory,
-            IModelMetadataProvider modelMetadataProvider)
-            : base(viewOptions, writerFactory, viewEngine, tempDataFactory, diagnosticSource, modelMetadataProvider)
-        {
-            if (loggerFactory == null)
-            {
-                throw new ArgumentNullException(nameof(loggerFactory));
-            }
-
-            Logger = loggerFactory.CreateLogger<PdfResultExecutor>();
-        }
-
-        /// <summary>
-        /// Gets the <see cref="ILogger"/>.
-        /// </summary>
-        protected ILogger Logger { get; }
-
-        /// <summary>
-        /// Attempts to find the <see cref="IView"/> associated with <paramref name="viewResult"/>.
-        /// </summary>
-        /// <param name="actionContext">The <see cref="ActionContext"/> associated with the current request.</param>
-        /// <param name="viewResult">The <see cref="ViewResult"/>.</param>
-        /// <returns>A <see cref="ViewEngineResult"/>.</returns>
-        public virtual ViewEngineResult FindView(ActionContext actionContext, PdfResult viewResult)
-        {
-            if (actionContext == null)
-            {
-                throw new ArgumentNullException(nameof(actionContext));
-            }
-
-            if (viewResult == null)
-            {
-                throw new ArgumentNullException(nameof(viewResult));
-            }
-
-            var viewEngine = viewResult.ViewEngine ?? ViewEngine;
-            var viewName = viewResult.ViewName ?? GetActionName(actionContext);
-
-            var result = viewEngine.GetView(executingFilePath: null, viewPath: viewName, isMainPage: false);
-            var originalResult = result;
-            if (!result.Success)
-            {
-                result = viewEngine.FindView(actionContext, viewName, isMainPage: false);
-            }
-
-            if (!result.Success)
-            {
-                if (originalResult.SearchedLocations.Any())
-                {
-                    if (result.SearchedLocations.Any())
-                    {
-                        // Return a new ViewEngineResult listing all searched locations.
-                        var locations = new List<string>(originalResult.SearchedLocations);
-                        locations.AddRange(result.SearchedLocations);
-                        result = ViewEngineResult.NotFound(viewName, locations);
-                    }
-                    else
-                    {
-                        // GetView() searched locations but FindView() did not. Use first ViewEngineResult.
-                        result = originalResult;
-                    }
-                }
-            }
-
-            if (result.Success)
-            {
-                if (DiagnosticSource.IsEnabled("Microsoft.AspNetCore.Mvc.ViewFound"))
-                {
-                    DiagnosticSource.Write(
-                        "Microsoft.AspNetCore.Mvc.ViewFound",
-                        new
-                        {
-                            actionContext = actionContext,
-                            isMainPage = false,
-                            result = viewResult,
-                            viewName = viewName,
-                            view = result.View,
-                        });
-                }
-            }
-            else
-            {
-                if (DiagnosticSource.IsEnabled("Microsoft.AspNetCore.Mvc.ViewNotFound"))
-                {
-                    DiagnosticSource.Write(
-                        "Microsoft.AspNetCore.Mvc.ViewNotFound",
-                        new
-                        {
-                            actionContext = actionContext,
-                            isMainPage = false,
-                            result = viewResult,
-                            viewName = viewName,
-                            searchedLocations = result.SearchedLocations
-                        });
-                }
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Executes the <see cref="IView"/> asynchronously.
-        /// </summary>
-        /// <param name="actionContext">The <see cref="ActionContext"/> associated with the current request.</param>
-        /// <param name="view">The <see cref="IView"/>.</param>
-        /// <param name="viewResult">The <see cref="ViewResult"/>.</param>
-        /// <returns>A <see cref="Task"/> which will complete when view execution is completed.</returns>
-        public async Task ExecuteAsync(ActionContext actionContext, IView view, PdfResult viewResult)
-        {
-            if (actionContext == null)
-            {
-                throw new ArgumentNullException(nameof(actionContext));
-            }
-
-            if (view == null)
-            {
-                throw new ArgumentNullException(nameof(view));
-            }
-
-            if (viewResult == null)
-            {
-                throw new ArgumentNullException(nameof(viewResult));
-            }
-
-            var response = actionContext.HttpContext.Response;
-
-            response.Headers.Add("Content-Disposition", viewResult.ContentDisposition);
-
-
-            string resolvedContentType = null;
-            Encoding resolvedContentTypeEncoding = null;
-            ResponseContentTypeHelper.ResolveContentTypeAndEncoding( viewResult.ContentType, viewResult.ContentType, DefaultContentType, out resolvedContentType, out resolvedContentTypeEncoding);
-
-            response.ContentType = resolvedContentType;
-            response.StatusCode = viewResult.StatusCode;
-
-            var razorStream = new MemoryStream();
-
-            var writer = new StreamWriter(razorStream);
-            var viewContext = new ViewContext(
-                actionContext,
-                view,
-                viewResult.ViewData,
-                viewResult.TempData,
-                writer,
-                ViewOptions.HtmlHelperOptions);
-
-            await view.RenderAsync(viewContext);
-            await writer.FlushAsync();
-            
-            Document document = new Document();
-            // step 2
-            var pdfWriter = PdfWriter.GetInstance(document, response.Body);
-
-            HTMLWorker worker = new HTMLWorker(document);
-
-            razorStream.Position = 0;
-            using (var tr = new StreamReader(razorStream))
-            {
-                // step 3
-                document.Open();
-                worker.StartDocument();
-                worker.Parse(tr);
-
-                worker.EndDocument();
-                worker.Close();
-                document.Close();
-            }
-            
-        }
-
-        private static string GetActionName(ActionContext context)
-        {
-            if (context == null)
-            {
-                throw new ArgumentNullException(nameof(context));
-            }
-
-            object routeValue;
-            if (!context.RouteData.Values.TryGetValue(ActionNameKey, out routeValue))
-            {
-                return null;
-            }
-
-            var actionDescriptor = context.ActionDescriptor;
-            string normalizedValue = null;
-            string value;
-            if (actionDescriptor.RouteValues.TryGetValue(ActionNameKey, out value) &&
-                !string.IsNullOrEmpty(value))
-            {
-                normalizedValue = value;
-            }
-
-            var stringRouteValue = routeValue?.ToString();
-            if (string.Equals(normalizedValue, stringRouteValue, StringComparison.OrdinalIgnoreCase))
-            {
-                return normalizedValue;
-            }
-
-            return stringRouteValue;
-        }
-    }
-}
