--- RazorPDF/RazorPDFCore.nuspec
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <title>$title$</title>
-    <authors>$author$</authors>
-    <owners>$author$</owners>
-    <licenseUrl>https://github.com/ole1986/RazorPDF</licenseUrl>
-    <projectUrl>https://github.com/ole1986/RazorPDF</projectUrl>
-    <iconUrl>https://github.com/ole1986/RazorPDF</iconUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>$description$</description>
-    <releaseNotes>Compatible with ASP.NET Core</releaseNotes>
-    <copyright>Copyright 2016</copyright>
-    <tags>razor pdf itextsharp</tags>
-  </metadata>
-</package>
